refactor(tga): split period_trends.rs into submodule under 500-line cap
period_trends.rs was 660 lines — over the 500-line hard cap. Split into:
  - report/period_trends/mod.rs   (20 lines) — thin facade, re-exports
  - report/period_trends/model.rs (57 lines) — AuthorPeriodSummary struct
  - report/period_trends/query.rs (425 lines) — query fn + all SQL helpers
  - report/period_trends/tests.rs (248 lines) — 9 unit tests

Public API paths (crate::report::{query_author_period_trends, AuthorPeriodSummary})
are unchanged. All 16 period-trend tests continue to pass.
