--- crates/trusty-git-analytics/src/report/period_trends.rs
@@ -1,660 +1,0 @@
-//! N-week period-trend rollup for the longitudinal contributor-profile epic (#558).
-//!
-//! Provides [`query_author_period_trends`] which aggregates existing per-week DB
-//! rows into fixed-width N-week windows for a single canonical author. No schema
-//! changes are required — all data is read from existing tables (`commits`,
-//! `authors`, `fact_commit_effort`, `classifications`, `pull_requests`).
-
-use std::collections::HashMap;
-
-use chrono::{Datelike, Duration, IsoWeek, NaiveDate};
-use rusqlite::params;
-use serde::{Deserialize, Serialize};
-
-use crate::core::db::Database;
-use crate::report::drilldown::{
-    extract_provider_logins, lookup_author_for_drilldown, query_effort_histogram, query_pr_metrics,
-    PrMetrics,
-};
-use crate::report::errors::{ReportError, Result};
-
-// ─── Data model ──────────────────────────────────────────────────────────────
-
-/// Aggregated per-author metrics for a single N-week period window.
-///
-/// Why: longitudinal contributor profiling (epic #558) needs period-level
-/// roll-ups of the existing per-week data so callers can plot trends over time
-/// without fetching the full weekly-activity grain.
-/// What: holds all dimensions required for one period window — label, date bounds,
-/// commit count, per-category breakdown, effort histogram, averaged quality score,
-/// ticket coverage percentage, PR metrics, and repositories touched. All fields
-/// derive from existing DB schema; no migrations required.
-/// Test: see `tests::period_trends_basic_windowing` and related tests below.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct AuthorPeriodSummary {
-    /// Human-readable period label, e.g. `"2026-W01..W04"`.
-    pub period_label: String,
-
-    /// Inclusive start of the period (ISO 8601 date, e.g. `"2026-01-05"`).
-    pub since: String,
-
-    /// Inclusive end of the period (ISO 8601 date, e.g. `"2026-02-01"`).
-    pub until: String,
-
-    /// Total commits in this period window.
-    pub commit_count: u64,
-
-    /// Per-category commit counts (`HashMap<category, count>`).
-    pub categories: HashMap<String, u64>,
-
-    /// Effort size → commit count for effort-scored commits in the window.
-    pub effort_histogram: HashMap<String, u32>,
-
-    /// Average quality score across all weeks in the window.
-    /// Computed as the mean of `fact_weekly_quality.quality_score` rows for
-    /// this author that fall within the window. `0.0` when no quality data is
-    /// available.
-    pub quality_score: f64,
-
-    /// Ticketed commits / total commits in `[0.0, 1.0]`.
-    /// `0.0` when `commit_count == 0`.
-    pub ticketed_pct: f64,
-
-    /// Aggregated PR metrics for this period (cycle-time, counts).
-    pub pr_metrics: PrMetrics,
-
-    /// Distinct repositories touched in this period.
-    pub repositories: Vec<String>,
-}
-
-// ─── Query implementation ─────────────────────────────────────────────────────
-
-/// Aggregate existing per-week data for one canonical author into N-week
-/// period windows.
-///
-/// Why: the contributor-profile epic (#558) requires trend data bucketed into
-/// multi-week periods (e.g. 4-week sprints) rather than raw per-week rows,
-/// enabling callers to render velocity trend lines and period-over-period
-/// comparisons without rebuilding aggregation logic.
-/// What: computes the set of ISO weeks for the author in `[since, until]`,
-/// partitions them into chunks of `window_weeks`, and for each chunk reuses
-/// `query_effort_histogram`, `query_pr_metrics`, and inline SQL against
-/// `commits` / `classifications` / `fact_weekly_quality` to assemble an
-/// [`AuthorPeriodSummary`]. Reads existing schema only — no migration needed.
-/// Returns an empty `Vec` when the author has no commits in the window.
-///
-/// # Parameters
-///
-/// - `db` — open database handle
-/// - `email` — canonical email matched case-insensitively
-/// - `window_weeks` — number of ISO weeks per period bucket (minimum 1)
-/// - `since` — optional ISO 8601 lower bound (inclusive); `None` = start of data
-/// - `until` — optional ISO 8601 upper bound (inclusive); `None` = end of data
-///
-/// # Errors
-///
-/// Returns [`ReportError`] on any DB failure.
-///
-/// Test: see `tests` module below.
-pub fn query_author_period_trends(
-    db: &Database,
-    email: &str,
-    window_weeks: u32,
-    since: Option<&str>,
-    until: Option<&str>,
-) -> Result<Vec<AuthorPeriodSummary>> {
-    let window_weeks = window_weeks.max(1);
-
-    // Resolve the author to get their aliases (needed for PR queries).
-    let author_row = lookup_author_for_drilldown(db, email)?;
-    let aliases_json = match author_row {
-        Some((_, _, _, ref aliases)) => aliases.clone(),
-        None => return Ok(Vec::new()),
-    };
-    let logins = extract_provider_logins(&aliases_json);
-
-    // Determine the effective date bounds from the commits table.
-    let (data_since, data_until) = effective_date_bounds(db, email, since, until)?;
-    let (data_since, data_until) = match (data_since, data_until) {
-        (Some(s), Some(u)) => (s, u),
-        _ => return Ok(Vec::new()), // no commits in scope
-    };
-
-    // Parse dates and enumerate ISO weeks in the range.
-    let start_date = parse_iso_date(&data_since)?;
-    let end_date = parse_iso_date(&data_until)?;
-
-    // Collect the sorted sequence of distinct ISO-week start dates covered by
-    // commits for this author. This avoids creating empty period buckets for
-    // calendar gaps where the author was inactive.
-    let week_starts = weeks_in_range(start_date, end_date);
-    if week_starts.is_empty() {
-        return Ok(Vec::new());
-    }
-
-    // Partition weeks into fixed-width buckets of `window_weeks`.
-    let mut summaries = Vec::new();
-    for chunk in week_starts.chunks(window_weeks as usize) {
-        let period_since_date = *chunk.first().expect("non-empty chunk");
-        let period_until_date = chunk
-            .last()
-            .expect("non-empty chunk")
-            .checked_add_signed(Duration::days(6))
-            .unwrap_or(*chunk.last().expect("non-empty chunk"));
-
-        let period_since = period_since_date.format("%Y-%m-%d").to_string();
-        let period_until = period_until_date.format("%Y-%m-%d").to_string();
-
-        let period_label = make_period_label(period_since_date, period_until_date);
-
-        let summary =
-            build_period_summary(db, email, &logins, period_label, period_since, period_until)?;
-        summaries.push(summary);
-    }
-
-    Ok(summaries)
-}
-
-// ─── Helpers ──────────────────────────────────────────────────────────────────
-
-/// Return `(min_timestamp, max_timestamp)` for commits by `email` in the
-/// optional `[since, until]` window.
-fn effective_date_bounds(
-    db: &Database,
-    email: &str,
-    since: Option<&str>,
-    until: Option<&str>,
-) -> Result<(Option<String>, Option<String>)> {
-    let conn = db.connection();
-    let mut stmt = conn
-        .prepare(
-            "SELECT MIN(c.timestamp), MAX(c.timestamp) \
-             FROM commits c \
-             JOIN authors a ON a.id = c.author_id \
-             WHERE LOWER(a.canonical_email) = LOWER(?1) \
-               AND (?2 IS NULL OR c.timestamp >= ?2) \
-               AND (?3 IS NULL OR c.timestamp <= ?3)",
-        )
-        .map_err(crate::core::TgaError::from)?;
-
-    let (min_ts, max_ts) = stmt
-        .query_row(params![email, since, until], |row| {
-            Ok((
-                row.get::<_, Option<String>>(0)?,
-                row.get::<_, Option<String>>(1)?,
-            ))
-        })
-        .map_err(crate::core::TgaError::from)?;
-
-    // Trim to YYYY-MM-DD.
-    let trim = |s: Option<String>| s.map(|v| v.get(..10).unwrap_or(&v).to_string());
-    Ok((trim(min_ts), trim(max_ts)))
-}
-
-/// Build one [`AuthorPeriodSummary`] for the given period bounds.
-fn build_period_summary(
-    db: &Database,
-    email: &str,
-    logins: &[String],
-    period_label: String,
-    period_since: String,
-    period_until: String,
-) -> Result<AuthorPeriodSummary> {
-    let conn = db.connection();
-
-    // Commit count + ticketed count + repositories.
-    let mut stmt = conn
-        .prepare(
-            "SELECT COUNT(*), \
-                    SUM(CASE WHEN c.ticketed = 1 THEN 1 ELSE 0 END) \
-             FROM commits c \
-             JOIN authors a ON a.id = c.author_id \
-             WHERE LOWER(a.canonical_email) = LOWER(?1) \
-               AND c.timestamp >= ?2 \
-               AND c.timestamp <= ?3 || 'T23:59:59Z'",
-        )
-        .map_err(crate::core::TgaError::from)?;
-
-    let (commit_count, ticketed_count): (u64, u64) = stmt
-        .query_row(params![email, period_since, period_until], |row| {
-            Ok((
-                row.get::<_, i64>(0)? as u64,
-                row.get::<_, Option<i64>>(1)?.unwrap_or(0) as u64,
-            ))
-        })
-        .map_err(crate::core::TgaError::from)?;
-
-    // Repositories.
-    let mut repo_stmt = conn
-        .prepare(
-            "SELECT DISTINCT c.repository \
-             FROM commits c \
-             JOIN authors a ON a.id = c.author_id \
-             WHERE LOWER(a.canonical_email) = LOWER(?1) \
-               AND c.timestamp >= ?2 \
-               AND c.timestamp <= ?3 || 'T23:59:59Z' \
-             ORDER BY c.repository",
-        )
-        .map_err(crate::core::TgaError::from)?;
-
-    let repo_rows = repo_stmt
-        .query_map(params![email, period_since, period_until], |row| {
-            row.get::<_, String>(0)
-        })
-        .map_err(crate::core::TgaError::from)?;
-
-    let mut repositories = Vec::new();
-    for r in repo_rows {
-        repositories.push(r.map_err(crate::core::TgaError::from)?);
-    }
-
-    // Per-category counts.
-    let mut cat_stmt = conn
-        .prepare(
-            "SELECT cl.category, COUNT(*) \
-             FROM commits c \
-             JOIN authors a ON a.id = c.author_id \
-             LEFT JOIN classifications cl ON cl.id = c.classification_id \
-             WHERE LOWER(a.canonical_email) = LOWER(?1) \
-               AND cl.category IS NOT NULL \
-               AND c.timestamp >= ?2 \
-               AND c.timestamp <= ?3 || 'T23:59:59Z' \
-             GROUP BY cl.category",
-        )
-        .map_err(crate::core::TgaError::from)?;
-
-    let cat_rows = cat_stmt
-        .query_map(params![email, period_since, period_until], |row| {
-            Ok((row.get::<_, String>(0)?, row.get::<_, i64>(1)?))
-        })
-        .map_err(crate::core::TgaError::from)?;
-
-    let mut categories: HashMap<String, u64> = HashMap::new();
-    for r in cat_rows {
-        let (cat, cnt) = r.map_err(crate::core::TgaError::from)?;
-        categories.insert(cat, cnt as u64);
-    }
-
-    // Effort histogram — reuse the existing helper.
-    let effort = query_effort_histogram(db, email, Some(&period_since), Some(&period_until))?;
-    let effort_histogram: HashMap<String, u32> = effort.histogram;
-
-    // Average quality score from fact_weekly_quality (if table exists & has rows).
-    let quality_score = query_avg_quality_score(db, email, &period_since, &period_until)?;
-
-    // PR metrics — reuse the existing helper with the period bounds.
-    let pr_metrics = query_pr_metrics(db, logins, Some(&period_since), Some(&period_until))?;
-
-    let ticketed_pct = if commit_count > 0 {
-        ticketed_count as f64 / commit_count as f64
-    } else {
-        0.0
-    };
-
-    Ok(AuthorPeriodSummary {
-        period_label,
-        since: period_since,
-        until: period_until,
-        commit_count,
-        categories,
-        effort_histogram,
-        quality_score,
-        ticketed_pct,
-        pr_metrics,
-        repositories,
-    })
-}
-
-/// Query the average `quality_score` from `fact_weekly_quality` for the author
-/// in the given period.  Returns `0.0` when the table has no matching rows.
-///
-/// Why: quality data is keyed by `(iso_year, iso_week)` integers; this helper
-/// converts the YYYY-MM-DD period bounds to those integers for comparison.
-///
-/// What: AVGs `quality_score` for rows whose `(iso_year, iso_week)` falls within
-/// the period bounds.
-///
-/// Test: covered indirectly — period_trends tests that seed quality rows exercise
-/// the non-zero path; tests without quality rows verify the `0.0` fallback.
-fn query_avg_quality_score(
-    db: &Database,
-    email: &str,
-    period_since: &str,
-    period_until: &str,
-) -> Result<f64> {
-    // Parse the ISO dates to extract iso_year and iso_week integers for the
-    // comparison. `fact_weekly_quality` stores (iso_year INTEGER, iso_week
-    // INTEGER) — NOT a text `week` column.
-    let since_date = parse_iso_date(period_since)?;
-    let until_date = parse_iso_date(period_until)?;
-
-    let since_iso = since_date.iso_week();
-    let until_iso = until_date.iso_week();
-
-    let since_year = since_iso.year();
-    let since_week = since_iso.week() as i32;
-    let until_year = until_iso.year();
-    let until_week = until_iso.week() as i32;
-
-    let conn = db.connection();
-    let result: rusqlite::Result<Option<f64>> = conn.query_row(
-        "SELECT AVG(fwq.quality_score) \
-         FROM fact_weekly_quality fwq \
-         WHERE LOWER(fwq.author_email) = LOWER(?1) \
-           AND (fwq.iso_year > ?2 OR (fwq.iso_year = ?2 AND fwq.iso_week >= ?3)) \
-           AND (fwq.iso_year < ?4 OR (fwq.iso_year = ?4 AND fwq.iso_week <= ?5))",
-        params![email, since_year, since_week, until_year, until_week],
-        |row| row.get(0),
-    );
-    match result {
-        Ok(Some(avg)) => Ok(avg),
-        Ok(None) => Ok(0.0),
-        Err(rusqlite::Error::SqliteFailure(_, _)) | Err(rusqlite::Error::QueryReturnedNoRows) => {
-            Ok(0.0)
-        }
-        Err(e) => Err(ReportError::Core(crate::core::TgaError::from(e))),
-    }
-}
-
-/// Parse a `YYYY-MM-DD` string into a `NaiveDate`.
-fn parse_iso_date(s: &str) -> Result<NaiveDate> {
-    NaiveDate::parse_from_str(s, "%Y-%m-%d")
-        .map_err(|e| ReportError::Report(format!("invalid date string '{s}': {e}")))
-}
-
-/// Return the sequence of Monday dates (ISO week starts) that cover the
-/// range `[start, end]` inclusive.
-fn weeks_in_range(start: NaiveDate, end: NaiveDate) -> Vec<NaiveDate> {
-    // Round `start` down to its ISO week's Monday.
-    let first_monday = iso_week_monday(start.iso_week());
-    let last_monday = iso_week_monday(end.iso_week());
-
-    let mut weeks = Vec::new();
-    let mut current = first_monday;
-    while current <= last_monday {
-        weeks.push(current);
-        current = current
-            .checked_add_signed(Duration::weeks(1))
-            .unwrap_or(current);
-        if current == weeks.last().copied().unwrap_or(current) {
-            break; // safety against infinite loop
-        }
-    }
-    weeks
-}
-
-/// Return the Monday of the given ISO week.
-fn iso_week_monday(isoweek: IsoWeek) -> NaiveDate {
-    // NaiveDate::from_isoywd_opt: year, week, weekday (Monday=1).
-    NaiveDate::from_isoywd_opt(isoweek.year(), isoweek.week(), chrono::Weekday::Mon)
-        .expect("valid ISO week always produces a valid Monday")
-}
-
-/// Build a human-readable period label, e.g. `"2026-W01..W04"` or
-/// `"2026-W52..2027-W02"` (cross-year).
-fn make_period_label(since: NaiveDate, until: NaiveDate) -> String {
-    let since_week = since.iso_week();
-    let until_week = until.iso_week();
-    if since_week.year() == until_week.year() {
-        format!(
-            "{}-W{:02}..W{:02}",
-            since_week.year(),
-            since_week.week(),
-            until_week.week()
-        )
-    } else {
-        format!(
-            "{}-W{:02}..{}-W{:02}",
-            since_week.year(),
-            since_week.week(),
-            until_week.year(),
-            until_week.week()
-        )
-    }
-}
-
-// ─── Tests ────────────────────────────────────────────────────────────────────
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::core::db::Database;
-    use rusqlite::params;
-
-    fn seed_author(db: &Database, name: &str, email: &str) -> i64 {
-        db.connection()
-            .execute(
-                "INSERT INTO authors (canonical_name, canonical_email, aliases) \
-                 VALUES (?1, ?2, '[]')",
-                params![name, email],
-            )
-            .expect("insert author");
-        db.connection().last_insert_rowid()
-    }
-
-    fn seed_commit(
-        db: &Database,
-        sha: &str,
-        author_id: i64,
-        timestamp: &str,
-        ticketed: i64,
-        category: Option<&str>,
-    ) {
-        // Insert classification if needed.
-        let cls_id: Option<i64> = if let Some(cat) = category {
-            db.connection()
-                .execute(
-                    "INSERT OR IGNORE INTO classifications \
-                     (id, category, confidence, method) \
-                     VALUES (NULL, ?1, 0.9, 'exact_rule')",
-                    params![cat],
-                )
-                .expect("insert classification");
-            let id: i64 = db
-                .connection()
-                .query_row(
-                    "SELECT id FROM classifications WHERE category = ?1 LIMIT 1",
-                    params![cat],
-                    |r| r.get(0),
-                )
-                .expect("get cls id");
-            Some(id)
-        } else {
-            None
-        };
-
-        db.connection()
-            .execute(
-                "INSERT INTO commits \
-                 (sha, author_id, author_name, author_email, timestamp, message, \
-                  repository, insertions, deletions, ticketed, classification_id) \
-                 VALUES (?1, ?2, 'n', 'e', ?3, 'm', 'repo-a', 5, 2, ?4, ?5)",
-                params![sha, author_id, timestamp, ticketed, cls_id],
-            )
-            .expect("insert commit");
-    }
-
-    /// Why: baseline test — two 4-week windows should be produced for 8 weeks
-    /// of commits, with correct commit counts per window.
-    /// What: seeds 8 weeks of commits, calls `query_author_period_trends` with
-    /// `window_weeks = 4`, asserts two periods with correct commit counts.
-    /// Test: this test itself.
-    #[test]
-    fn period_trends_basic_windowing() {
-        let db = Database::open_in_memory().expect("open");
-        let aid = seed_author(&db, "Alice", "alice@example.com");
-
-        // 8 commits, one per week starting 2024-W01.
-        let weeks = [
-            "2024-01-01T00:00:00Z", // W01
-            "2024-01-08T00:00:00Z", // W02
-            "2024-01-15T00:00:00Z", // W03
-            "2024-01-22T00:00:00Z", // W04
-            "2024-01-29T00:00:00Z", // W05
-            "2024-02-05T00:00:00Z", // W06
-            "2024-02-12T00:00:00Z", // W07
-            "2024-02-19T00:00:00Z", // W08
-        ];
-        for (i, ts) in weeks.iter().enumerate() {
-            seed_commit(&db, &format!("sha{i}"), aid, ts, 0, None);
-        }
-
-        let trends =
-            query_author_period_trends(&db, "alice@example.com", 4, None, None).expect("query");
-
-        assert_eq!(trends.len(), 2, "expected 2 period windows for 8 weeks");
-        assert_eq!(
-            trends[0].commit_count, 4,
-            "first window: 4 commits, got {}",
-            trends[0].commit_count
-        );
-        assert_eq!(
-            trends[1].commit_count, 4,
-            "second window: 4 commits, got {}",
-            trends[1].commit_count
-        );
-    }
-
-    /// Why: period labels must follow the `"YYYY-Www..Www"` format and the
-    /// since/until fields must be valid ISO dates.
-    /// What: calls `query_author_period_trends`, inspects the first period's
-    /// metadata fields.
-    /// Test: this test itself.
-    #[test]
-    fn period_trends_label_and_date_format() {
-        let db = Database::open_in_memory().expect("open");
-        let aid = seed_author(&db, "Bob", "bob@example.com");
-        seed_commit(&db, "sha1", aid, "2024-01-01T00:00:00Z", 0, None);
-        seed_commit(&db, "sha2", aid, "2024-01-08T00:00:00Z", 0, None);
-
-        let trends =
-            query_author_period_trends(&db, "bob@example.com", 4, None, None).expect("query");
-        assert!(!trends.is_empty());
-        let p = &trends[0];
-        assert!(
-            p.period_label.contains("-W"),
-            "period_label must contain '-W': {}",
-            p.period_label
-        );
-        assert_eq!(p.since.len(), 10, "since must be YYYY-MM-DD: {}", p.since);
-        assert_eq!(p.until.len(), 10, "until must be YYYY-MM-DD: {}", p.until);
-    }
-
-    /// Why: ticketed_pct must reflect the fraction of ticketed commits correctly.
-    /// What: seeds 4 commits (2 ticketed), asserts `ticketed_pct ≈ 0.5`.
-    /// Test: this test itself.
-    #[test]
-    fn period_trends_ticketed_pct() {
-        let db = Database::open_in_memory().expect("open");
-        let aid = seed_author(&db, "Carol", "carol@example.com");
-        seed_commit(&db, "s1", aid, "2024-01-01T00:00:00Z", 1, None);
-        seed_commit(&db, "s2", aid, "2024-01-02T00:00:00Z", 1, None);
-        seed_commit(&db, "s3", aid, "2024-01-03T00:00:00Z", 0, None);
-        seed_commit(&db, "s4", aid, "2024-01-04T00:00:00Z", 0, None);
-
-        let trends =
-            query_author_period_trends(&db, "carol@example.com", 4, None, None).expect("query");
-        assert!(!trends.is_empty());
-        let pct = trends[0].ticketed_pct;
-        assert!(
-            (pct - 0.5).abs() < 1e-9,
-            "ticketed_pct should be 0.5, got {pct}"
-        );
-    }
-
-    /// Why: an author with no commits must return an empty Vec, not an error.
-    /// What: seeds an author but no commits, asserts the result is empty.
-    /// Test: this test itself.
-    #[test]
-    fn period_trends_empty_for_no_commits() {
-        let db = Database::open_in_memory().expect("open");
-        seed_author(&db, "Dave", "dave@example.com");
-        let trends =
-            query_author_period_trends(&db, "dave@example.com", 4, None, None).expect("query");
-        assert!(trends.is_empty(), "no commits → empty Vec, got: {trends:?}");
-    }
-
-    /// Why: an unknown email must return an empty Vec (not an error), consistent
-    /// with the design contract.
-    /// What: queries for an email not in `authors`, asserts empty result.
-    /// Test: this test itself.
-    #[test]
-    fn period_trends_empty_for_unknown_email() {
-        let db = Database::open_in_memory().expect("open");
-        let trends =
-            query_author_period_trends(&db, "nobody@example.com", 4, None, None).expect("query");
-        assert!(trends.is_empty(), "unknown email → empty Vec");
-    }
-
-    /// Why: categories must be aggregated correctly per window.
-    /// What: seeds 3 feature + 1 bugfix commits in one window, asserts the
-    /// `categories` map reflects those counts.
-    /// Test: this test itself.
-    #[test]
-    fn period_trends_category_aggregation() {
-        let db = Database::open_in_memory().expect("open");
-        let aid = seed_author(&db, "Eve", "eve@example.com");
-        seed_commit(&db, "e1", aid, "2024-01-01T00:00:00Z", 0, Some("feature"));
-        seed_commit(&db, "e2", aid, "2024-01-02T00:00:00Z", 0, Some("feature"));
-        seed_commit(&db, "e3", aid, "2024-01-03T00:00:00Z", 0, Some("feature"));
-        seed_commit(&db, "e4", aid, "2024-01-04T00:00:00Z", 0, Some("bugfix"));
-
-        let trends =
-            query_author_period_trends(&db, "eve@example.com", 4, None, None).expect("query");
-        assert!(!trends.is_empty());
-        let cats = &trends[0].categories;
-        assert_eq!(
-            cats.get("feature").copied(),
-            Some(3),
-            "expected 3 feature commits"
-        );
-        assert_eq!(
-            cats.get("bugfix").copied(),
-            Some(1),
-            "expected 1 bugfix commit"
-        );
-    }
-
-    /// Why: the `since`/`until` filter must constrain the data to the requested
-    /// date range only.
-    /// What: seeds commits spanning two months; queries with a 1-month filter;
-    /// asserts only commits in that range are counted.
-    /// Test: this test itself.
-    #[test]
-    fn period_trends_date_filter() {
-        let db = Database::open_in_memory().expect("open");
-        let aid = seed_author(&db, "Frank", "frank@example.com");
-        seed_commit(&db, "f1", aid, "2024-01-10T00:00:00Z", 0, None);
-        seed_commit(&db, "f2", aid, "2024-01-20T00:00:00Z", 0, None);
-        seed_commit(&db, "f3", aid, "2024-02-10T00:00:00Z", 0, None); // out of window
-
-        let trends = query_author_period_trends(
-            &db,
-            "frank@example.com",
-            4,
-            Some("2024-01-01"),
-            Some("2024-01-31"),
-        )
-        .expect("query");
-
-        let total: u64 = trends.iter().map(|p| p.commit_count).sum();
-        assert_eq!(
-            total, 2,
-            "filter to Jan 2024 should yield 2 commits, got {total}"
-        );
-    }
-
-    /// Why: the helper `weeks_in_range` must produce the correct count and order.
-    /// What: passes a known 4-week range, asserts 4 Mondays are returned.
-    /// Test: this test itself.
-    #[test]
-    fn weeks_in_range_produces_correct_mondays() {
-        let start = NaiveDate::from_ymd_opt(2024, 1, 1).expect("date");
-        let end = NaiveDate::from_ymd_opt(2024, 1, 28).expect("date");
-        let weeks = weeks_in_range(start, end);
-        assert_eq!(weeks.len(), 4, "should be 4 Mondays in the range");
-        for w in &weeks {
-            assert_eq!(w.weekday(), chrono::Weekday::Mon, "{w} must be a Monday");
-        }
-    }
-}
